requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
idna==3.4
importlib-metadata==4.13.0
markdown-it-py==2.2.0
mdurl==0.1.2
mypy-extensions==0.4.4
nornir==3.3.0
nornir-utils==0.2.0
ply==3.11
pycryptodomex==3.17
Pygments==2.15.1
pysnmp-pyasn1==1.1.3
pysnmp-pysmi==1.1.10
pysnmplib==5.0.21
requests==2.29.0
rich==13.3.5
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
typing-extensions==4.5.0
urllib3==1.26.15
zipp==3.15.0
